requirements.txt
gradio>=4.28
streamlit>=1.34
pandas>=2.2
numpy>=1.26
matplotlib>=3.8
shap>=0.45
scikit-learn>=1.4
joblib>=1.3
Pillow>=10.2
opencv-python>=4.9
torch>=2.2
torchvision>=0.17
python-dotenv>=1.0
openai>=1.35
